Add files via upload
This repository contains a Shiny application that visualizes historical newspaper data from Britain and Ireland on an interactive map. Users can filter the displayed data by selecting a range of publication years and specific countries. The application leverages the power of R's leaflet library for mapping and tidyverse for data manipulation, providing an intuitive and responsive interface. Ideal for historians, researchers, or anyone interested in exploring the geographical and temporal distribution of historical newspapers.